project structure: flask application (WIP)
