+++ stellarisdashboard/dashboard_app/static/style.css
