+++ stellarisdashboard/dashboard_app/static/toggle_visibility.js
