+++ stellarisdashboard/dashboard_app/templates/game_index.html
