+++ stellarisdashboard/dashboard_app/templates/layout.html
